# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Real Estate API",
        "description": "API to manage real estate properties like houses and apartments.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "support@realestateapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get a list of all properties",
                "operationId": "ae50890d9da51a5cfe01cd2151d4077a",
                "responses": {
                    "200": {
                        "description": "A list of properties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Property"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Properties"
                ],
                "summary": "Create a new property",
                "operationId": "debc241247e30f0490b9e236483217d0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Property"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Property created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/properties/{id}": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get a specific property",
                "operationId": "1ffeff79c18b3281ff5e01339a5831e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the property to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the property",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Properties"
                ],
                "summary": "Update a specific property",
                "operationId": "d5f20f7ec2d5b865985ee2a307481311",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the property to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Property"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Property updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Properties"
                ],
                "summary": "Delete a specific property",
                "operationId": "a1f3cd4975f784df9ae3f810e13a2066",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the property to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Property deleted successfully"
                    }
                }
            }
        },
        "/api/properties/search": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Search properties based on filters",
                "operationId": "04a80dc4ab88389047430a1611931979",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of the property",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address of the property",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of the property",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bedrooms",
                        "in": "query",
                        "description": "Number of bedrooms",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Maximum price",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of filtered properties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Property"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/properties/nearby": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Find properties within a given radius of a geographic point",
                "operationId": "627fb293df7aaa0737902363ec639fa9",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Latitude of the location",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Longitude of the location",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Search radius in kilometers",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of properties within the radius",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Property"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Property": {
                "required": [
                    "type",
                    "address",
                    "price",
                    "size",
                    "bedrooms"
                ],
                "properties": {
                    "type": {
                        "description": "The type of property (e.g., apartment, house)",
                        "type": "string"
                    },
                    "address": {
                        "description": "The address of the property",
                        "type": "string"
                    },
                    "price": {
                        "description": "The price of the property",
                        "type": "integer"
                    },
                    "size": {
                        "description": "The size of the property in square meters",
                        "type": "integer"
                    },
                    "bedrooms": {
                        "description": "The number of bedrooms in the property",
                        "type": "integer"
                    },
                    "latitude": {
                        "description": "The latitude coordinate of the property",
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "description": "The longitude coordinate of the property",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_key_security_example": {
                "type": "apiKey",
                "description": "A short description for the security scheme",
                "name": "api_key",
                "in": "header"
            },
            "oauth2_security_example": {
                "type": "oauth2",
                "description": "OAuth2 security scheme description",
                "flow": "implicit",
                "authorizationUrl": "http://example.com/auth",
                "scopes": {
                    "read:projects": "read your projects",
                    "write:projects": "modify projects"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Properties",
            "description": "Properties"
        }
    ]
}